Refactored chat button binding
