+++ betterrolls5e/module.json
@@ -48,7 +48,7 @@
 		"./scripts/betterrolls5e.js",
 		"./scripts/custom-roll.js",
 		"./scripts/init.js",
-		"./scripts/chat-damage-buttons.js",
+		"./scripts/chat-message.js",
 		"./scripts/utils.js",
 		"./scripts/hooks.js",
 		"./scripts/item-tab.js"
