+++ betterrolls5e/templates/red-damage-overlay.html
@@ -1,0 +1,14 @@
+<span class="dmgBtn-container-br temporary">
+    <button data-modifier="1">
+        <i class="fas fa-user-minus" title="{{ localize "br5e.chat.damageButtons.fullDamage.hint" }}"></i>
+    </button>
+    <button data-modifier="0.5">
+        <i class="fas fa-user-shield" title="{{ localize "br5e.chat.damageButtons.halfDamage.hint" }}"></i>
+    </button>
+    <button data-modifier="2">
+        <i class="fas fa-user-injured" title="{{ localize "br5e.chat.damageButtons.doubleDamage.hint" }}"></i>
+    </button>
+    <button data-modifier="-1">
+        <i class="fas fa-user-plus" title="{{ localize "br5e.chat.damageButtons.healing.hint" }}"></i>
+    </button>
+</span>
